+++ src/smt/smt_engine.cpp
@@ -91,6 +91,7 @@
 #include "theory/quantifiers/ce_guided_instantiation.h"
 #include "theory/quantifiers/fun_def_process.h"
 #include "theory/quantifiers/macros.h"
+#include "theory/quantifiers/single_inv_partition.h"
 #include "theory/quantifiers/quantifiers_rewriter.h"
 #include "theory/quantifiers/term_util.h"
 #include "theory/sort_inference.h"
