+++ src/theory/quantifiers/anti_skolem.h
@@ -24,7 +24,7 @@
 #include "expr/type_node.h"
 #include "context/cdhashset.h"
 #include "context/cdo.h"
-#include "theory/quantifiers/ce_guided_single_inv.h"
+#include "theory/quantifiers/single_inv_partition.h"
 #include "theory/quantifiers_engine.h"
 
 namespace CVC4 {
